Fix loading of default config file
If the default config file is badly formed tfsec will not fail, instead it will continue to dereference a nil tfsecConfig pointer and crash.

To avoid this, first check if the default config file exists, if not we'll continue on happily; otherwise load the config and fail if there are any errors.