Fix AWS051 `value is null` warning
When `aws_rds_cluster` attribute `kms_key_id` was explicitly set to
`null`, tfsec did not catch the disabled encryption and issued a
warning:

> WARNING: skipped AWS051 check due to error(s): value is null

The added test case reproduces this issue.

The test case does not catch the issue though, as the test execution
logic seems to differ from the real scanning logic, in that it doesn't
try to format the result.