Add a flag that allows passed checks to be output.
Adds a boolean property to the result list that is set whenever a check passes or fails. Respects the concise-output in the default formatter to not print statistics, but adds the total number of checks, passing, and failing if the flag is provided.

Updates all of the existing checks to return a passing result rather than null.